% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.file.R
\docType{data}
\name{paper.file}
\alias{paper.file}
\title{Text of Hebert paper Introduction
P.D.N. Hebert, E.H. Penton, J.M. Burns, D.H. Janzen, W. Hallwachs, Ten species in one: DNA barcoding reveals cryptic species in the neotropical skipper butterfly Astraptes fulgerator, Proc. Natl. Acad. Sci. USA 101 (2004) 14812-14817.}
\format{
A text file
}
\source{
\url{https://www.pnas.org/content/101/41/14812}
}
\usage{
paper.file
}
\description{
Text of Hebert paper Introduction
P.D.N. Hebert, E.H. Penton, J.M. Burns, D.H. Janzen, W. Hallwachs, Ten species in one: DNA barcoding reveals cryptic species in the neotropical skipper butterfly Astraptes fulgerator, Proc. Natl. Acad. Sci. USA 101 (2004) 14812-14817.
}
\keyword{datasets}
