% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_match_first.R
\name{rw_match_first}
\alias{rw_match_first}
\title{Match first sentence to paragraphs (rw_match_first)}
\usage{
rw_match_first(
  infile,
  title = "title",
  part = "section",
  project = "name",
  parabreak = "\\r\\n"
)
}
\arguments{
\item{infile}{A text file containing several paragraphs.}

\item{title}{Title of text (e.g. paper reference).}

\item{part}{Part of the text (e.g. paper section).
Used in filenames.}

\item{project}{Project nickname for use in filenames}

\item{parabreak}{The characters that designate paragraph
breaks in the input}
}
\value{
A text file of paragraphs with first sentences
removed and random list of first sentences
}
\description{
Input is a text file with multiple paragraphs.
Output is a text file for each paragraph with the first
sentence removed and with the first sentences listed
in random order.
}
