% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_rand_para.R
\name{rw_rand_para}
\alias{rw_rand_para}
\title{Randomize paragraphs (rw_rand_para)}
\usage{
rw_rand_para(
  infile,
  title = "title",
  part = "section",
  project = "name",
  parabreak = "\\r\\n"
)
}
\arguments{
\item{infile}{A text file containing several paragraphs.}

\item{title}{Title of text (e.g. paper reference).}

\item{part}{Part of the text (e.g. paper section).
Used in filenames.}

\item{project}{Project nickname for use in filenames}

\item{parabreak}{The characters that designate paragraph
breaks in the input}
}
\value{
A text file with paragraphs in random order
}
\description{
Input is a text file with multiple paragraphs.
Output is a text file with paragraph order
randomized.
}
