% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_remove_first.R
\name{rw_remove_first}
\alias{rw_remove_first}
\title{Remove first sentence of paragraph (rw_remove_first)}
\usage{
rw_remove_first(
  infile,
  title = "title",
  part = "section",
  project = "name",
  parabreak = "\\r\\n"
)
}
\arguments{
\item{infile}{A text file containing several paragraphs.}

\item{title}{Title of text (e.g. paper reference).}

\item{part}{Part of the text (e.g. paper section).
Used in filenames.}

\item{project}{Project nickname for use in filenames}

\item{parabreak}{The characters that designate paragraph
breaks in the input}
}
\value{
A text file with paragraphs without first sentence
}
\description{
Input is a text file with multiple paragraphs.
Output is a text file for each paragraph with the first
sentence removed.
}
